+++ completions/ccs.bash
@@ -17,7 +17,7 @@ _ccs() {
         cword=$COMP_CWORD
     fi
 
-    local commands="init set preset use verify doctor ls list current show rm remove help --help --version -h -v"
+    local commands="init set preset use slim verify doctor ls list current show rm remove help --help --version -h -v"
     local providers_dir="${CCS_DIR:-$HOME/.config/ccs}/providers"
     local -a providers=()
     if [ -d "$providers_dir" ]; then
@@ -55,6 +55,10 @@ _ccs() {
                 # shellcheck disable=SC2207
                 COMPREPLY=( $(compgen -W "$providers_str" -- "$cur") )
             fi
+            ;;
+        slim)
+            # shellcheck disable=SC2207
+            COMPREPLY=( $(compgen -W "--apply" -- "$cur") )
             ;;
         show)
             if [ "$cword" -eq 2 ]; then